requirements.txt
fastapi
uvicorn[standard]
gunicorn
pandas
numpy
scikit-learn
joblib
python-multipart
azure-storage-blob
